+++ src/main/kotlin/blackjack/domain/model/CardNumber.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain.model
 
 @JvmInline
 value class CardNumber(
