+++ src/main/kotlin/blackjack/domain/model/Deck.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain.model
 
 class Deck {
     companion object {
