+++ src/main/kotlin/blackjack/domain/model/Suit.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain.model
 
 @JvmInline
 value class Suit(
