+++ src/test/kotlin/blackjack/domain/model/DeckTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
